requirements.txt
apache-airflow==2.3.0
boto3==1.24.0
psycopg2==2.9.3
pandas==1.4.0
matplotlib==3.5.1
seaborn==0.11.2
requests==2.27.1
SQLAlchemy==1.4.31
jupyter==1.0.0
great-expectations==0.15.4
pytest==7.0.1
